+++ workspace/app/services/dev-settings.svc.js
@@ -54,31 +54,14 @@ angular.module('impacWorkspace').service('DevSettings', function ($q, ImpacRoute
     return _self._defaults = angular.copy(DEFAULTS);
   }
 
-  // NOTE: this will be refactored a lot when mnoe routes are added as default
-  // in impac-angular.
   // TODO: add routes settings to #/settings, 'basic' & 'advance' tabs could be good.
   this.buildRoutesConfig = function (mnoeUrl, impacUrl, multipleWatchableMode) {
     var routesConfig = {
       mnoHub: mnoeUrl.host + mnoeUrl.base,
-      impacPrefix: "/impac",
       impacApi: impacUrl.host + impacUrl.base,
-      dashboards: {
-        index: mnoeUrl.host + mnoeUrl.base + "/impac/dashboards"
-      },
-      widgets: {
-        index: mnoeUrl.host + mnoeUrl.base + "/impac/widgets",
-        create: mnoeUrl.host + mnoeUrl.base + "/impac/dashboards/:dashboard_id/widgets"
-      },
       kpis: {
+        // MnoEnterprise KPI templates proxy endpoint
         index: mnoeUrl.host + mnoeUrl.base + "/impac/kpis",
-        create: mnoeUrl.host + mnoeUrl.base + "/impac/dashboards/:dashboard_id/kpis",
-        update: mnoeUrl.host + mnoeUrl.base + "/impac/kpis/:id",
-        del: mnoeUrl.host + mnoeUrl.base + "/impac/kpis/:id"
-      },
-      alerts: {
-        index: mnoeUrl.host + mnoeUrl.base + "/impac/alerts",
-        create: mnoeUrl.host + mnoeUrl.base + "/impac/kpis/:kpi_id/alerts",
-        del: mnoeUrl.host + mnoeUrl.base + "/impac/alerts/:id"
       }
     }
     // Removes custom index path, defaulting kpi discovery to impac api, where KPIs with